TAC-199: ShipStation - when removing failed OrderLabels we now check that they've actually been saved first, otherwise a NotFound is thrown.
We generally only persist the OrderLabels once we've successfully got the label from the courier.
We do sometimes save it early though e.g. if we had to fetch rates first.

+++ library/CG/ShipStation/Carrier/Label/Creator/Other.php
@@ -318,7 +318,9 @@ class Other implements CreatorInterface, LoggerAwareInterface
     protected function removeFailedOrderLabel(OrderLabel $orderLabel, array $errorMsgs): void
     {
         $this->logNotice('Failed to generate label for Order %s, reason(s): %s', [$orderLabel->getOrderId(), str_replace('%', '%%', implode('; ', $errorMsgs))], [static::LOG_CODE, 'Fail']);
-        $this->orderLabelService->remove($orderLabel);
+        if ($orderLabel->getId()) {
+            $this->orderLabelService->remove($orderLabel);
+        }
     }
 
     /**
